Add Azure blob journal storage narration comments
